+++ CHANGELOG.md
@@ -1,6 +1,22 @@
 # Release Notes for Stripe for Craft Commerce
 
+## 2.2.0 - 2019-10-23
+
+### Fixed
+- Fixed a bug where Stripe webhook reponses could contain unnecessary JavaScript. ([##86](https://github.com/craftcms/commerce-stripe/issues/86))  
+
 ## 2.1.2 - 2019-09-04
+
+### Added
+- Added support for resolving subscription billing issues.
+- Added `craft\commerce\stripe\models\forms\Subscription`.
+- Added `craft\commerce\stripe\models\forms\SwitchPlans::$billingCycleAnchor`.
+- Added `craft\commerce\stripe\services\Invoices::getUserInvoices()`.
+
+### Changed
+- Update Craft Commerce requirements to require Craft Commerce 2.2.
+- Improved support for specifying trial when starting a subscription.
+- Improved support for specifying billing cycle changes when switching subscription plans.
 
 ### Fixed
 - Fixed a bug where payment form errors were not being caught and returned. ([#75](https://github.com/craftcms/commerce-stripe/issues/75))
@@ -11,6 +27,7 @@
 ### Fixed
 - Fixed a bug where payment forms would sometimes throw an error. ([#59](https://github.com/craftcms/commerce-stripe/issues/59))
 - Fixed a PHP error when retrieving the subscriptions next payment amount. ([#71](https://github.com/craftcms/commerce-stripe/issues/71))
+- Fixed a bug where failed 3DS authentications would have no error code and error message.
 
 ## 2.1.0 - 2019-07-24